refactor(block): log error on data verification failed
