+++ block/manager_test.go
@@ -410,106 +410,6 @@ func TestGetDataSignature_NilSigner(t *testing.T) {
 	require.ErrorContains(err, "signer is nil; cannot sign data")
 }
 
-// TestIsValidSignedData covers valid, nil, wrong proposer, and invalid signature cases for isValidSignedData.
-func TestIsValidSignedData(t *testing.T) {
-	require := require.New(t)
-	privKey, _, err := crypto.GenerateKeyPair(crypto.Ed25519, 256)
-	require.NoError(err)
-	testSigner, err := noopsigner.NewNoopSigner(privKey)
-	require.NoError(err)
-	proposerAddr, err := testSigner.GetAddress()
-	require.NoError(err)
-	gen := genesispkg.NewGenesis(
-		"testchain",
-		1,
-		time.Now(),
-		proposerAddr,
-	)
-	m := &Manager{
-		signer:  testSigner,
-		genesis: gen,
-	}
-
-	t.Run("valid signed data", func(t *testing.T) {
-		batch := &types.Data{
-			Txs: types.Txs{types.Tx("tx1"), types.Tx("tx2")},
-		}
-		sig, err := m.getDataSignature(batch)
-		require.NoError(err)
-		pubKey, err := m.signer.GetPublic()
-		require.NoError(err)
-		signedData := &types.SignedData{
-			Data:      *batch,
-			Signature: sig,
-			Signer: types.Signer{
-				PubKey:  pubKey,
-				Address: proposerAddr,
-			},
-		}
-		assert.True(t, m.isValidSignedData(signedData))
-	})
-
-	t.Run("nil signed data", func(t *testing.T) {
-		assert.False(t, m.isValidSignedData(nil))
-	})
-
-	t.Run("nil Txs", func(t *testing.T) {
-		signedData := &types.SignedData{
-			Data: types.Data{},
-			Signer: types.Signer{
-				Address: proposerAddr,
-			},
-		}
-		signedData.Txs = nil
-		assert.False(t, m.isValidSignedData(signedData))
-	})
-
-	t.Run("wrong proposer address", func(t *testing.T) {
-		batch := &types.Data{
-			Txs: types.Txs{types.Tx("tx1")},
-		}
-		sig, err := m.getDataSignature(batch)
-		require.NoError(err)
-		pubKey, err := m.signer.GetPublic()
-		require.NoError(err)
-		wrongAddr := make([]byte, len(proposerAddr))
-		copy(wrongAddr, proposerAddr)
-		wrongAddr[0] ^= 0xFF // flip a bit
-		signedData := &types.SignedData{
-			Data:      *batch,
-			Signature: sig,
-			Signer: types.Signer{
-				PubKey:  pubKey,
-				Address: wrongAddr,
-			},
-		}
-		assert.False(t, m.isValidSignedData(signedData))
-	})
-
-	t.Run("invalid signature", func(t *testing.T) {
-		batch := &types.Data{
-			Txs: types.Txs{types.Tx("tx1")},
-		}
-		sig, err := m.getDataSignature(batch)
-		require.NoError(err)
-		pubKey, err := m.signer.GetPublic()
-		require.NoError(err)
-		// Corrupt the signature
-		badSig := make([]byte, len(sig))
-		copy(badSig, sig)
-		badSig[0] ^= 0xFF
-		signedData := &types.SignedData{
-			Data:      *batch,
-			Signature: badSig,
-			Signer: types.Signer{
-				PubKey:  pubKey,
-				Address: proposerAddr,
-			},
-		}
-		assert.False(t, m.isValidSignedData(signedData))
-	})
-}
-
 // TestManager_execValidate tests the execValidate method for various header/data/state conditions.
 func TestManager_execValidate(t *testing.T) {
 	require := require.New(t)
